Only force CouchDB to use the updated local.ini file if it is not bound to the correct port

+++ cookbooks/couchdb/recipes/default.rb
@@ -21,6 +21,7 @@ package "couchdb" do
   action :install
 end
 
+# Bind CouchDB to the correct port so that it can be accessed from the host
 cookbook_file "/etc/couchdb/local.ini" do
   source "local.ini"
   owner "couchdb"
@@ -28,10 +29,12 @@ cookbook_file "/etc/couchdb/local.ini" do
   mode 0664
 end
 
-# Force CouchDB to use the updated local.ini file. Kill the CouchDB Erlang
-# process rather than restarting the service as a) restarting the service
-# doesn't seem to reload the local.ini file and b) restarting the service
-# immediately after installation seems to fail.
+# Force CouchDB to use the updated local.ini file if it is not bound to the
+# correct port. Kill the CouchDB Erlang process rather than restarting the
+# service as a) restarting the service doesn't seem to reload the local.ini
+# file and b) restarting the service immediately after installation seems to
+# fail.
 execute "kill-couchdb" do
   command "kill -9 `cat /var/run/couchdb/couchdb.pid`"
+  only_if "netstat -an | grep '127\.0\.0\.1:5984'"
 end
